Added sidebar through layout + connected admin panel to side bar
