+++ src/app/(sidebar)/admin/recipes/page.tsx
@@ -13,10 +13,9 @@ export default async function RecipesManagement() {
         .select()
 
     return (
-        <div className="flex min-h-screen bg-white">
 
-            {/* Main content */}
-            <div className="flex-1 p-8">
+
+            <div className="flex-1">
                 <div className="mb-8">
                     <div className="flex items-center gap-2">
                         <Link href="/admin" className="text-[#2e6930] hover:underline">
@@ -50,7 +49,7 @@ export default async function RecipesManagement() {
                 </div>
 
             </div>
-        </div>
+
     )
 }
