+++ src/app/(sidebar)/page.tsx
@@ -3,7 +3,7 @@ import Dashboard from "./dashboard/page";
 export default async function Home() {
 	return (
 		<div>
-			<Dashboard></Dashboard>
+			<Dashboard />
 		</div>
 	);
 }
